+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Fixed removed collections.Mapping usage with python 3.10
 - Added splunkd full service check during SHC member add. Did run into timeout when having lots of apps.
 - Updated URL for downloads of splunk install archives
 - Follow symlinks during Splunk archive extraction